supabase-deployment/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==23.0.0
Werkzeug==3.0.3